Add 'wrong' quality.
There are too many cases where you need to invent stuff because the
original is either untranslatable, use words not part of any dictionnary
or requires many research to actually get the meaning.

'bad' do not begin to describe it, so add the 'wrong' quality, which is
worse than bad, but does not imply that the translated text is badly
formulated.

It can also be used to mark back translations that turns out to be wrong
after being unknown.