Add support for completing nothing into an entire screen.
This is an alternative to prefilled text for cases where there is the
possibility that the result is predicable.  Pressing tab with an empty
input will yield the configured 'complete completion'.

This was already the case if the shown translations only consisted of
one unique word, in which case, the word completion would complete it
right away by hitting tab on an empty string.