Add a sophisticated check to report strange variable references.
The aim is to catch errors like translating "Bergen Village" to
\v[area.heat-village.name].  If a translation contains a variable
reference and the original does not contain the same variable reference
or the equivalent original text, then issue a notice.

There are loads of possible false positives (e.g. "Autumn Rise" vs
"Autumn's Rise" in the original text), but it's still useful.