checker.py: Make font size checker account for the size of icons.
This may not be useful right now, but it will soon enough™.

Of course, the metrics for icons will need to be specified, but this
only concerns one icon for now, and two in the future. I can even tell
you their size: "tech" has width 17, and rumors says a new icon named
"ascended" will have width 24.  Both are for the normal font.
Icons can now be specified in "extra_metrics", as if there were
characters.