packfile: Make 'split' give more information.
Curating map_file by hand is a good way to reduce the size of git
repositories.  Using loads of small files is great for deduplication,
but worse for blob compression.  But using a single big file can easily
take a lot of space because of the lack of deduplication, until git
repacks the blobs with delta compression.