common: Support walking more degenerate packs with GameWalker.
Previously, the program would crash when trying to handle degenerate
cases, such as stale texts, but it turns out that some commands require
this, such as "./packfile.py filter".

Now, it is even possible, even if not recommended, to filter things that
are not pack files, but are still indexed by file_dict_path_str.  This
even includes string caches.